+++ src/app/shared/components/headline-banner/headline-banner.component.ts
@@ -1,0 +1,9 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'fs-headline-banner',
+  templateUrl: './headline-banner.component.html',
+  styleUrls: ['./headline-banner.component.scss']
+})
+export class HeadlineBannerComponent {
+}