+++ src/app/shared/shared.module.ts
@@ -3,6 +3,7 @@ import { CommonModule } from '@angular/common';
 import { NgbModule } from '@ng-bootstrap/ng-bootstrap';
 import { RouterModule } from '@angular/router';
 import { RoundButtonComponent } from 'src/app/shared/components/round-button/fs-round-button.component';
+import { HeadlineBannerComponent } from 'src/app/shared/components/headline-banner/headline-banner.component';
 import { OfferTileComponent } from 'src/app/shared/components/offer-tile/offer-tile.component';
 import { ErrorBackTextComponent } from 'src/app/shared/components/error-back-text/error-back-text.component';
 import { LoadingSpinnerComponent } from 'src/app/shared/components/loading-spinner/loading-spinner.component';
@@ -36,7 +37,7 @@ import {
   faVideo,
   faPlusCircle,
   faEdit,
-  faEye
+  faEye,
 } from '@fortawesome/free-solid-svg-icons';
 import { CheckPermissionsDirective } from './directives/check-permissions.directive';
 
@@ -51,6 +52,7 @@ import { CheckPermissionsDirective } from './directives/check-permissions.direct
     SmallOfferTileComponent,
     NavBackDirective,
     CheckPermissionsDirective,
+    HeadlineBannerComponent,
   ],
   imports: [CommonModule, NgbModule, RouterModule, FontAwesomeModule],
   exports: [
@@ -66,6 +68,7 @@ import { CheckPermissionsDirective } from './directives/check-permissions.direct
     SmallOfferTileComponent,
     NavBackDirective,
     CheckPermissionsDirective,
+    HeadlineBannerComponent,
   ],
 })
 export class SharedModule {